% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhytoIn-package.R
\docType{package}
\name{PhytoIn-package}
\alias{PhytoIn}
\alias{PhytoIn-package}
\title{PhytoIn: Tools for vegetation analysis and forest inventory}
\description{
The package provides functions and datasets for phytosociological analysis,
forest inventory, biomass/carbon estimation, volume stratification, and
visualization (see \code{\link{phytoparam}}, \code{\link{AGB}},
\code{\link{stratvol}}, \code{\link{BAplot}}, \code{\link{plot.param}},
\code{\link{collector.curve}}, \code{\link{rarefaction}},
\code{\link{stats}}, \code{\link{summary.param}}).
}
\section{Datasets}{

\itemize{
  \item \code{\link{quadrat.df}}
  \item \code{\link{point.df}}
  \item \code{\link{quadrat2_plot.df}}
  \item \code{\link{quadrat2_tree.df}}
  \item \code{\link{quadrat3_rect.df}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PhytoIn/PhytoIn}
  \item Report bugs at \url{https://github.com/PhytoIn/PhytoIn/issues}
}

}
\author{
\strong{Maintainer}: Rodrigo Augusto Santinelo Pereira \email{raspereira@usp.br}

}
\keyword{internal}
