% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quadrat2_plot.df}
\alias{quadrat2_plot.df}
\title{Quadrat dataset with coordinates for basal area plotting}
\format{
A data frame with 219 rows and 6 variables:
\describe{
  \item{Plot}{Quadrat identifier (e.g., X1Y1).}
  \item{CBH}{Circumference at breast height (cm).}
  \item{h}{Total height of the tree (m).}
  \item{Species}{Scientific name of the species.}
  \item{x}{X coordinate (m) of the lower-left corner of the quadrat.}
  \item{y}{Y coordinate (m) of the lower-left corner of the quadrat.}
}
}
\usage{
data(quadrat2_plot.df)
}
\description{
A dataset containing tree measurements from 25 quadrats (5 × 5 m each) 
sampled in a forest fragment located at the campus of the University of São Paulo, 
Ribeirão Preto, Brazil. This dataset is organized to demonstrate the use of the 
function \code{BAplot()} for visualizing basal areas. In addition to tree measurements, 
it provides the spatial coordinates of the lower-left corner of each quadrat.
}
\keyword{datasets}
