% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quadrat2_tree.df}
\alias{quadrat2_tree.df}
\title{Quadrat dataset with simulated individual tree coordinates}
\format{
A data frame with 219 rows and 7 variables:
\describe{
  \item{Plot}{Quadrat identifier (e.g., X1Y1).}
  \item{CBH}{Circumference at breast height (cm).}
  \item{h}{Total height of the tree (m).}
  \item{Species}{Scientific name of the species.}
  \item{x}{Simulated X coordinate (m) of the tree inside the quadrat.}
  \item{y}{Simulated Y coordinate (m) of the tree inside the quadrat.}
}
}
\usage{
data(quadrat2_tree.df)
}
\description{
A dataset containing tree measurements from 25 quadrats (5 × 5 m each)
sampled in a forest fragment located at the campus of the University of Sao Paulo,
Ribeirao Preto, Brazil. This dataset corresponds to the same trees as in
\code{quadrat2_plot.df}, but includes simulated coordinates (in meters) for
each individual tree inside the quadrats. It is organized to demonstrate the
use of the function \code{BAplot()} with the argument \code{ind.coord = TRUE}.
}
\keyword{datasets}
